tornado-api/requirements.txt
aiofiles==0.4.0
aiohttp==3.4.4
aiomysql==0.0.19
bcrypt==3.1.4
demjson==2.2.4
jwt==0.5.4
lxml==3.7.2
peewee==2.10.2
peewee-async==0.5.12
PyExecJS==1.5.1
PyJWT==1.6.4
PyMySQL==0.9.2
pyOpenSSL==18.0.0
redis==2.10.5
requests==2.13.0
tornado==5.1.1
WTForms==2.2.1
WTForms-JSON==0.3.3
wtforms-tornado==0.0.2
